+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [6.2.0] - 2024-01-29
+
+### Added
+- Deprecated soft-dependency packages are detected and a warning with upgrade instructions is shown in the console.
+
+### Changed
+- Support for Draco 3D Data Compression is now provided by [*Draco for Unity* (com.unity.cloud.draco)][DracoForUnity], which is a fork of and replaces [*DracoUnity* (com.atteneder.draco)][DracoUnity].
+
+### Fixed
+- Compiler error when Newtonsoft JSON package was not installed.
+- All Draco vertex attributes are assigned by identifier instead of type. As a result, tangents are now decoded properly instead of recalculated.
+- Compilation error when scripting define `GLTFAST_BUILTIN_SHADER_GRAPH` is set.
+- `GltfImport.IsTextureYFlipped` returns correct result for non-KTX textures.
+
 ## [6.1.0] - 2024-01-17
 
 ### Added
@@ -1062,6 +1076,7 @@ This release contains multiple breaking changes. Please read the [upgrade guide]
 [Entities1.0]: https://docs.unity3d.com/Packages/com.unity.entities@1.0
 [KtxUnity]: https://github.com/atteneder/KtxUnity
 [DanDovi]: https://github.com/DanDovi
+[DracoForUnity]: https://docs.unity3d.com/Packages/com.unity.cloud.draco@latest
 [DracoUnity]: https://github.com/atteneder/DracoUnity
 [aurorahcx]: https://github.com/aurorahcx
 [Battlehub0x]: https://github.com/Battlehub0x